requirements.txt
torch>=1.12.0
torchvision>=0.13.0
streamlit>=1.18.0
pillow>=9.0.1
pandas>=1.4.0
tqdm>=4.64.0
matplotlib>=3.5.0
